requirements.txt
rapidfuzz>=3.0.0
google-cloud-storage>=2.10.0
functions-framework>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0 